requirements.txt
# ── Data Generator ──
faker>=28.0.0
kafka-python>=2.0.2
orjson>=3.10.0
pyyaml>=6.0.1
click>=8.1.7

# ── Spark Jobs ──
pyspark==3.5.5
delta-spark==3.3.0

# ── Delta Viewer ──
deltalake>=0.18.0
pyarrow>=16.0.0
tabulate>=0.9.0
rich>=13.7.0

# ── Cleanup ──
humanize>=4.9.0

# ── Dashboard ──
streamlit>=1.35.0
duckdb>=0.10.0
